requirements.txt
numpy
scipy
tensorflow
huggingface_hub
pandas
matplotlib
datasets
scikit-learn
python-dotenv
torch
pyro-ppl
contextily
